Low level function to compute the pairwise genotype probabilities

This should never really be used directly by users. Rather, use the pairwise_genotype_probs() function that calls this.

Usage

pgp_rcpp(IXG, AF, isD, AD, debug)

Arguments

IXG

an integer matrix of the intersected intervals with columns

kIdx: 0-based index of the kid

pIdx: 0-based index of candidate parent

cIdx: 0-based index of chromosome

anck: ternary trit giving ancestry of kid at this segment

ancp: ternary trit giving ancestry of candidate parent at this segment

start: integer base pair of the start of the segment. Will probably be used just to know when we are reaching a new segment.

stop: integer base pair of the stop of the segment (probably won't be used)

mIdx: 0-based index of the marker

pos: base-pair position of the marker (likely will not be used)

gk: count of 1 alleles in the kid genotype (0, 1, or 2, or -1 for missing data)

gp: count of 1 alleles in the parent genotype (0, 1, or 2, or -1 for missing data)

AF

a numeric matrix with number of columns equal to the number of ancestries and ordered like A, B, C,... and number of rows equal to the number of variable and diagnostic markers combined.

isD

an integer vector of 0s and 1s with length equal to the number of variable and diagnostic markers combined. A 1 means the marker is a species-diagnostic marker and a 0 means otherwise.

AD

a numeric matrix with number of columns equal to the number of ancestries and number of rows equal to the total number of individuals that got put into the integer representation. Each row sums to one. These are the admixture fractions.

debug

an IntegerVector, but we only use the first element. debug(0) == 0 means no debug information. Otherwise the function returns a list with a lot of extra information that can be used to verify that the calculations are being done correctly. If debug(0) == 1, then information is returned at the rate of one row per segment, and no genotype information gets returned. If debug(0) == 2, then information is returned at one row per marker, and it includes the genotype probabilities for each marker in the unrelated case. If debug(0) == 3 then information is returned with one row for each marker and each distinct value of the ancestry of the segregated haplotype.