Collect inputs for calculating genotype pair probabilities

The purpose of this is to collect the necessary inputs and prepare everything to be passed into another function that will do the intersecting of intervals and then the calculation of genotype pair probabilities.

Usage

prepare_for_gpp(V, M, E, diag_err_rate = 0.005)

Arguments

V

Long tibble of variable-marker genotypes with columns chrom, pos, indiv, n.

M

long tibble of diagnostic markers with columns diag_spp, chrom, pos, indiv, n

E

The tibble of ancestral segments as it comes out of extend_ancestral_segments_3().

diag_err_rate

the rate at which diagnostic alleles might not be so diagnostic for a given species. Defaults to 1 in 200.

Value

This function retuns a list with a number of components. This list serves as one of the inputs to pairwise_genotype_probs(). The components of the return list are:

mKey: a tibble showing the integer equivalent of chrom-key + "-" position for indexing the markers.

AF: A tibble of allele frequencies.

“

Details

The things done by this function are:

Estimating allele frequencies of the variable markers in the different backgrounds, using an EM algorithm.

Estimating admixture fractions of the individuals using the extended ancestral segments in each of them.

Creating an integer representation of the extended segments that can be used for efficiently intersecting segments.

The V and M inputs are largely the "raw" long format inputs, and E is what comes out of extend_ancestral_segments_3().